# swagger_server/models/banda.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.tecnologia import Tecnologia  # noqa: F401,E501
from swagger_server import util


class Banda(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, velocidade: str=None, tecnologia: Tecnologia=None):  # noqa: E501
        """Banda - a model defined in Swagger

        :param id: The id of this Banda.  # noqa: E501
        :type id: int
        :param velocidade: The velocidade of this Banda.  # noqa: E501
        :type velocidade: str
        :param tecnologia: The tecnologia of this Banda.  # noqa: E501
        :type tecnologia: Tecnologia
        """
        self.swagger_types = {
            'id': int,
            'velocidade': str,
            'tecnologia': Tecnologia
        }

        self.attribute_map = {
            'id': 'id',
            'velocidade': 'velocidade',
            'tecnologia': 'tecnologia'
        }
        self._id = id
        self._velocidade = velocidade
        self._tecnologia = tecnologia

    @classmethod
    def from_dict(cls, dikt) -> 'Banda':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Banda of this Banda.  # noqa: E501
        :rtype: Banda
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Banda.


        :return: The id of this Banda.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Banda.


        :param id: The id of this Banda.
        :type id: int
        """

        self._id = id

    @property
    def velocidade(self) -> str:
        """Gets the velocidade of this Banda.


        :return: The velocidade of this Banda.
        :rtype: str
        """
        return self._velocidade

    @velocidade.setter
    def velocidade(self, velocidade: str):
        """Sets the velocidade of this Banda.


        :param velocidade: The velocidade of this Banda.
        :type velocidade: str
        """

        self._velocidade = velocidade

    @property
    def tecnologia(self) -> Tecnologia:
        """Gets the tecnologia of this Banda.


        :return: The tecnologia of this Banda.
        :rtype: Tecnologia
        """
        return self._tecnologia

    @tecnologia.setter
    def tecnologia(self, tecnologia: Tecnologia):
        """Sets the tecnologia of this Banda.


        :param tecnologia: The tecnologia of this Banda.
        :type tecnologia: Tecnologia
        """

        self._tecnologia = tecnologia