# swagger_server/models/retorno_banda.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.banda import Banda  # noqa: F401,E501
from swagger_server.models.retorno import Retorno  # noqa: F401,E501
from swagger_server import util


class RetornoBanda(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, codigo: int=None, descricao: str=None, registros: int=None, pagina: int=None, qtde_pagina: int=None, lista_bandas: List[Banda]=None):  # noqa: E501
        """RetornoBanda - a model defined in Swagger

        :param codigo: The codigo of this RetornoBanda.  # noqa: E501
        :type codigo: int
        :param descricao: The descricao of this RetornoBanda.  # noqa: E501
        :type descricao: str
        :param registros: The registros of this RetornoBanda.  # noqa: E501
        :type registros: int
        :param pagina: The pagina of this RetornoBanda.  # noqa: E501
        :type pagina: int
        :param qtde_pagina: The qtde_pagina of this RetornoBanda.  # noqa: E501
        :type qtde_pagina: int
        :param lista_bandas: The lista_bandas of this RetornoBanda.  # noqa: E501
        :type lista_bandas: List[Banda]
        """
        self.swagger_types = {
            'codigo': int,
            'descricao': str,
            'registros': int,
            'pagina': int,
            'qtde_pagina': int,
            'lista_bandas': List[Banda]
        }

        self.attribute_map = {
            'codigo': 'codigo',
            'descricao': 'descricao',
            'registros': 'registros',
            'pagina': 'pagina',
            'qtde_pagina': 'qtdePagina',
            'lista_bandas': 'listaBandas'
        }
        self._codigo = codigo
        self._descricao = descricao
        self._registros = registros
        self._pagina = pagina
        self._qtde_pagina = qtde_pagina
        self._lista_bandas = lista_bandas

    @classmethod
    def from_dict(cls, dikt) -> 'RetornoBanda':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The RetornoBanda of this RetornoBanda.  # noqa: E501
        :rtype: RetornoBanda
        """
        return util.deserialize_model(dikt, cls)

    @property
    def codigo(self) -> int:
        """Gets the codigo of this RetornoBanda.


        :return: The codigo of this RetornoBanda.
        :rtype: int
        """
        return self._codigo

    @codigo.setter
    def codigo(self, codigo: int):
        """Sets the codigo of this RetornoBanda.


        :param codigo: The codigo of this RetornoBanda.
        :type codigo: int
        """

        self._codigo = codigo

    @property
    def descricao(self) -> str:
        """Gets the descricao of this RetornoBanda.


        :return: The descricao of this RetornoBanda.
        :rtype: str
        """
        return self._descricao

    @descricao.setter
    def descricao(self, descricao: str):
        """Sets the descricao of this RetornoBanda.


        :param descricao: The descricao of this RetornoBanda.
        :type descricao: str
        """

        self._descricao = descricao

    @property
    def registros(self) -> int:
        """Gets the registros of this RetornoBanda.


        :return: The registros of this RetornoBanda.
        :rtype: int
        """
        return self._registros

    @registros.setter
    def registros(self, registros: int):
        """Sets the registros of this RetornoBanda.


        :param registros: The registros of this RetornoBanda.
        :type registros: int
        """

        self._registros = registros

    @property
    def pagina(self) -> int:
        """Gets the pagina of this RetornoBanda.


        :return: The pagina of this RetornoBanda.
        :rtype: int
        """
        return self._pagina

    @pagina.setter
    def pagina(self, pagina: int):
        """Sets the pagina of this RetornoBanda.


        :param pagina: The pagina of this RetornoBanda.
        :type pagina: int
        """

        self._pagina = pagina

    @property
    def qtde_pagina(self) -> int:
        """Gets the qtde_pagina of this RetornoBanda.


        :return: The qtde_pagina of this RetornoBanda.
        :rtype: int
        """
        return self._qtde_pagina

    @qtde_pagina.setter
    def qtde_pagina(self, qtde_pagina: int):
        """Sets the qtde_pagina of this RetornoBanda.


        :param qtde_pagina: The qtde_pagina of this RetornoBanda.
        :type qtde_pagina: int
        """

        self._qtde_pagina = qtde_pagina

    @property
    def lista_bandas(self) -> List[Banda]:
        """Gets the lista_bandas of this RetornoBanda.


        :return: The lista_bandas of this RetornoBanda.
        :rtype: List[Banda]
        """
        return self._lista_bandas

    @lista_bandas.setter
    def lista_bandas(self, lista_bandas: List[Banda]):
        """Sets the lista_bandas of this RetornoBanda.


        :param lista_bandas: The lista_bandas of this RetornoBanda.
        :type lista_bandas: List[Banda]
        """

        self._lista_bandas = lista_bandas