# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: BandaPython
  description: Controle de Banda Larga
  version: 1.0.0
servers:
- url: http://localhost:8086/bandas
tags:
- name: Banda Larga
paths:
  /banda:
    get:
      tags:
      - Banda Larga
      summary: Buscar Banda Larga
      description: Buscar Banda Larga.
      operationId: buscar_banda
      parameters:
      - name: velocidade
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: tecnologia
        in: query
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/Tecnologia'
      - name: pagina
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: integer
      - name: qtdePagina
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Retorno com sucesso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
        "500":
          description: Erro no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
      x-swagger-router-controller: BandaLarga
      x-openapi-router-controller: swagger_server.controllers.banda_larga_controller
    post:
      tags:
      - Banda Larga
      summary: Cadastrar Banda Larga
      description: Cadastrar Banda Larga.
      operationId: cadastrar_banda
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Banda'
        required: true
      responses:
        "200":
          description: Cadastrado com sucesso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
        "500":
          description: Erro no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
      x-swagger-router-controller: BandaLarga
      x-openapi-router-controller: swagger_server.controllers.banda_larga_controller
  /banda/{codigo}:
    put:
      tags:
      - Banda Larga
      summary: Alterar Banda Larga
      description: Alterar Banda Larga.
      operationId: alterar_banda
      parameters:
      - name: codigo
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Banda'
        required: true
      responses:
        "200":
          description: Alteração com sucesso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
        "500":
          description: Erro no servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
      x-swagger-router-controller: BandaLarga
      x-openapi-router-controller: swagger_server.controllers.banda_larga_controller
    delete:
      tags:
      - Banda Larga
      summary: Excluir Banda Larga
      description: Excluir Banda Larga.
      operationId: excluir_banda
      parameters:
      - name: codigo
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Exclusão com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
        "500":
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetornoBanda'
      x-swagger-router-controller: BandaLarga
      x-openapi-router-controller: swagger_server.controllers.banda_larga_controller
components:
  schemas:
    Retorno:
      type: object
      properties:
        codigo:
          type: integer
          format: int64
          example: 234
        descricao:
          type: string
        registros:
          type: integer
          example: 20
        pagina:
          type: integer
          example: 0
        qtdePagina:
          type: integer
          example: 15
    RetornoBanda:
      allOf:
      - $ref: '#/components/schemas/Retorno'
      - type: object
        properties:
          listaBandas:
            type: array
            items:
              $ref: '#/components/schemas/Banda'
    Banda:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 234
        velocidade:
          type: string
          example: "41"
        tecnologia:
          $ref: '#/components/schemas/Tecnologia'
    RetornoHistorico:
      allOf:
      - $ref: '#/components/schemas/Retorno'
      - type: object
        properties:
          listaHistoricos:
            type: array
            items:
              $ref: '#/components/schemas/Historico'
    Historico:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 234
        data:
          type: string
          format: date-time
        tipo:
          $ref: '#/components/schemas/TipoHistorico'
        sistema:
          $ref: '#/components/schemas/Sistema'
        descricao:
          type: string
    TipoHistorico:
      type: string
      enum:
      - post
      - put
      - delete
      - get
    Sistema:
      type: string
      enum:
      - banda
      - banda
      - tv
      - historico
    Tecnologia:
      type: string
      enum:
      - metalico
      - gpon